Move code from filter.go to readers.go and options.go
